+++ resources/views/components/menu/menu-children.blade.php
@@ -1,5 +1,0 @@
-@props(['children'])
-
-@foreach ($children as $child)
-    <a href="{{ url($child->url) }}"> {{ $child->name }} </a>
-@endforeach